Main Done

+++ Movie_Web/Areas/Admin/Controllers/AdminMoviesController.cs
@@ -118,7 +118,7 @@ namespace Movie_Web.Areas.Admin.Controllers
                 
                 movie.Alias = Utilities.SEOUrl(movie.MovieName) + "-" + Utilities.SEOUrl(Convert.ToString(movie.Episode));
                 movie.DirectorAlias = Utilities.SEOUrl(movie.Director);
-                var movie1 = _context.Movies.FirstOrDefault(x => x.Alias ==  movie.Alias);
+                var movie1 = _context.Movies.FirstOrDefault(x => x.Alias ==  movie.Alias && x.Episode == movie.Episode );
                 if (movie1 == null)
                 {
                     if (fImage != null)
@@ -141,7 +141,7 @@ namespace Movie_Web.Areas.Admin.Controllers
                 }
                 else
                 {
-                    _notifyService.Error("Movie is exist", 2);
+                    _notifyService.Error("Movie đã tồn tại, Hãy xóa movie này trc khi thêm", 2);
                 }
                
                 return RedirectToAction(nameof(Index));
@@ -479,6 +479,8 @@ namespace Movie_Web.Areas.Admin.Controllers
             var movie = await _context.Movies
                 .Include(m => m.Country)
                 .Include(m => m.Type)
+                .Include(m=> m.Actors)
+                .Include(m=> m.Categories)
                 .FirstOrDefaultAsync(m => m.MovieId == id);
             if (movie == null)
             {
@@ -510,9 +512,18 @@ namespace Movie_Web.Areas.Admin.Controllers
 
 
                 string Path1 = _environment.WebRootPath + "/images/MovieImage/" + movie.Image;
-                System.IO.File.Delete(Path1);
+                if(System.IO.File.Exists(Path1))
+                {
+                    
+                    System.IO.File.Delete(Path1);
+                }
+               
                 string Path2 = _environment.WebRootPath + "/images/MovieImageSlider/" + movie.ImageSlider;
-                System.IO.File.Delete(Path2);
+                if (System.IO.File.Exists(Path2))
+                {
+                    System.IO.File.Delete(Path2);
+                }
+
